+++ boxsdk/object/folder.py
@@ -124,98 +124,72 @@ class Folder(Item):
         return self._get_accelerator_upload_url()
 
     @api_call
-    def get_items(self, limit, offset=0, fields=None):
-        """Get the items in a folder.
+    def get_items(self, limit=None, offset=0, marker=None, use_marker=False, sort=None, direction=None, fields=None):
+        """
+        Get the items in a folder.
 
         :param limit:
             The maximum number of items to return.
         :type limit:
-            `int`
+            `int` or None
         :param offset:
-            The index at which to start returning items.
+            The index at which to start returning items when using offset-based pagin.
         :type offset:
             `int`
+        :param use_marker:
+            Whether to use marker-based paging instead of offset-based paging, defaults to False.
+        :type use_marker:
+            `bool`
+        :param marker:
+            The paging marker to start returning items from when using marker-based paging.
+        :type marker:
+            `unicode` or None
+        :param sort:
+            Item field to sort results on: 'id', 'name', or 'date'.
+        :type sort':
+            `unicode` or None
+        :param direction:
+            Sort direction for the items returned.
+        :type direction:
+            `unicode` or None
         :param fields:
             List of fields to request.
         :type fields:
             `Iterable` of `unicode`
         :returns:
-            A list of items in the folder.
+            The collection of items in the folder.
         :rtype:
-            `list` of :class:`Item`
+            `Iterable` of :class:`Item`
         """
         url = self.get_url('items')
-        params = {
-            'limit': limit,
-            'offset': offset,
-        }
-        if fields:
-            params['fields'] = ','.join(fields)
-        box_response = self._session.get(url, params=params)
-        response = box_response.json()
-        return [self.translator.translate(item['type'])(self._session, item['id'], item) for item in response['entries']]
+        additional_params = {}
+        if limit is not None:
+            additional_params['limit'] = limit
+        if sort:
+            additional_params['sort'] = sort
+        if direction:
+            additional_params['direction'] = direction
 
-    @api_call
-    def get_items_limit_offset(self, limit=None, offset=0, fields=None):
-        """
-        Get the items in a folder using limit-offset paging.
+        if use_marker:
+            additional_params['usemarker'] = True
+            return MarkerBasedObjectCollection(
+                url=url,
+                session=self._session,
+                limit=limit,
+                marker=marker,
+                fields=fields,
+                additional_params=additional_params,
+                return_full_pages=False,
+            )
 
-        :param limit:
-            The maximum number of items to return per page. If not specified, then will use the server-side default.
-        :type limit:
-            `int` or None
-        :param offset:
-            The index at which to start returning items.
-        :type offset:
-            `int`
-        :param fields:
-            List of fields to request.
-        :type fields:
-            `Iterable` of `unicode`
-        :returns:
-            An iterator of the items in the folder.
-        :rtype:
-            :class:`BoxObjectCollection`
-        """
         return LimitOffsetBasedObjectCollection(
-            self.session,
-            self.get_url('items'),
+            url=url,
+            session=self._session,
             limit=limit,
-            fields=fields,
             offset=offset,
-            return_full_pages=False,
-        )
-
-    @api_call
-    def get_items_marker(self, limit=None, marker=None, fields=None):
-        """
-        Get the items in a folder using marker-based paging.
-
-        :param limit:
-            The maximum number of items to return per page. If not specified, then will use the server-side default.
-        :type limit:
-            `int` or None
-        :param marker:
-            The offset index to start paging from.
-        :type marker:
-            `str` or None
-        :param fields:
-            List of fields to request.
-        :type fields:
-            `Iterable` of `unicode`
-        :returns:
-            An iterator of the items in the folder.
-        :rtype:
-            :class:`BoxObjectCollection`
-        """
-        return MarkerBasedObjectCollection(
-            self.session,
-            self.get_url('items'),
-            limit=limit,
             fields=fields,
-            marker=marker,
+            additional_params=additional_params,
             return_full_pages=False,
-            supports_limit_offset_paging=True,
         )
 
     @api_call
@@ -440,6 +414,45 @@ class Folder(Item):
             session=self._session,
             object_id=collab_id,
             response_object=collaboration_response,
+        )
+
+    def create_web_link(self, target_url, name=None, description=None):
+        """
+        Create a WebLink with a given url.
+
+        :param target_url:
+            The url the web link points to.
+        :type target_url:
+            `unicode`
+        :param name:
+            The name of the web link. Optional, the API will give it a default if not specified.
+        :type name:
+            `unicode` or None
+        :param description:
+            Description of the web link
+        :type name:
+            `unicode` or None
+        :return:
+            A :class:`WebLink` object.
+        :rtype:
+            :class:`WebLink`
+        """
+        url = self._session.get_url('web_links')
+        web_link_attributes = {
+            'url': target_url,
+            'parent': {
+                'id': self.object_id
+            }
+        }
+        if name is not None:
+            web_link_attributes['name'] = name
+        if description is not None:
+            web_link_attributes['description'] = description
+        response = self._session.post(url, data=json.dumps(web_link_attributes)).json()
+        return self.translator.translate(response['type'])(
+            session=self._session,
+            object_id=response['id'],
+            response_object=response
         )
 
     @api_call