+++ boxsdk/object/legal_hold_policy_assignment.py
@@ -1,0 +1,9 @@
+# coding:utf-8
+from __future__ import unicode_literals, absolute_import
+from .base_object import BaseObject
+
+
+class LegalHoldPolicyAssignment(BaseObject):
+
+    """Represents the legal hold policy assignment"""
+    _item_type = 'legal_hold_policy_assignment'