+++ test/unit/pagination/test_limit_offset_based_object_collection.py
@@ -1,14 +1,16 @@
 # coding: utf-8
 
 from __future__ import unicode_literals, absolute_import
+
 import json
+
 from mock import Mock, PropertyMock
 import pytest
 
 from boxsdk.network.default_network import DefaultNetworkResponse
 from boxsdk.pagination.limit_offset_based_object_collection import LimitOffsetBasedObjectCollection
 from boxsdk.session.box_response import BoxResponse
-from boxsdk.session.box_session import BoxSession
+from boxsdk.session.session import Session
 from .box_object_collection_test_base import BoxObjectCollectionTestBase
 
 
@@ -37,7 +39,7 @@ class TestLimitOffsetBasedObjectCollection(BoxObjectCollectionTestBase):
     @pytest.fixture()
     def mock_session(self, translator, mock_items_response):
         """Baseclass override."""
-        mock_box_session = Mock(BoxSession)
+        mock_box_session = Mock(Session)
         type(mock_box_session).translator = PropertyMock(
             return_value=translator
         )